requirements.txt
Flask==2.3.3
matplotlib==3.7.1
mysql-connector-python==8.0.33
gunicorn==20.2.0